update internal db on change

+++ sql/1_create_fort.sql
@@ -4,8 +4,8 @@ CREATE TABLE IF NOT EXISTS `sw_fort` (
   # 0 = unknown, 10 = pogo, 20 = ingress, 30 = lightship
   `type_id` tinyint NOT NULL,
   # 0 = unknown, 10 = gym, 15 = pokestop, 20 = portal, 30 = lighship_poi
-  `lat` float NOT NULL,
-  `lon` float NOT NULL,
+  `lat` double NOT NULL,
+  `lon` double NOT NULL,
   `name` varchar(128) DEFAULT NULL,
   `description` text DEFAULT NULL,
   `cover_image` varchar(256) DEFAULT NULL,
